Prevent threading issues when leaving rooms
A "BotContext" class is now passed to the commands and listeners,
instead of the "Bot" class itself.  This is to prevent commands from
leaving rooms while inside of a command class, which could cause the
"Bot" class's list of joined rooms to become out of sync with the
"StackoverflowChat" class's list of joined rooms.  The bot would leave
the room, but then, the command class's "onMessage" method returns a
response message, which gets sent to the room that the bot just left,
and which essentially causes the bot to re-join the room.

This commit also resolves a potential issue where the bot could leave a
chat room while a message for that room is sitting in the message send
queue. The bot would essentially rejoin the room once the message was
popped off the queue and sent.