Fix image URL regex
Before, the regex only checked if message began with "http" and ended
with an image file extension.  So, a message like "http://google.com foo
bar.jpg" would incorrectly match. This commit makes sure it only matches
if there are no spaces in the message (which would interrupt the URL).