Don't let Markdown formatting get in the way of parsing messages
The bot will now ignore Markdown formatting when parsing commands.  For
example, if the name of a command is bolded, it will ignore the bold
formatting and focus on the actual text content. Also works with
multi-line and fixed-font messages.

Also, the "learn" command is now aware of formatting. Any formatting the
user includes in a learn command's output is now properly preserved.
This includes multi-lined messages and fixed-font messages.