Fix issue with URLs in javadoc command
If a method is defined in a super class, and the user invokes the
javadoc command on a child class, then it will build the URL based on
the child class. That URL will not work because the method documentation
is not listed in the child class's javadocs. Instead, link to the super
class's javadoc page.

Example: /javadoc JFrame#getLocationOnScreen

The "getLocationOnScreen" method is defined in the Component class,
which is a super class of JFrame.  So, if it links to the JFrame javadoc
page, the method's javadoc will not be shown. Instead, the URL should
point to the javadoc page of the Component class.