Improve the way that the output of learned commands are recorded
Instead of having the bot convert the HTML-encoded message to Markdown
itself, use the SO Chat API to retrieve the raw, original chat message
that the user typed, and use that to get the learned command output.

Having the bot convert the HTML to Markdown itself gives inaccurate
results for certain edge cases.

https://github.com/JavaChat/OakBot/issues/36