Fix an issue that prevented the bot from joining any chat rooms
The cookie settings that are assigned to the HttpClient when calling
HttpClients.createDefault() were preventing the client from parsing the
cookies it receives when logging into the StackOverflow account. As a
consequence, the bot could not join any chat rooms because it didn't
have the cookies with the login tokens.

The solution was to explicitly set the cookie spec that the client uses.

See: https://stackoverflow.com/q/36473478/13379