Change the way message replies are handled
SO Chat now lets you specify the parent ID of the message you are
replying to in the HTTP POST parameters. This is what happens when you
click "reply". Previously, the ID had to be
specified at the beginning of the chat message. Monospace-formatted
replies render properly when using the parameter.